Cells put values with proper ID
When I changed the cell-id shape from a map to an integer, I
erroneously changed the value put part of the cell impl. I was using
the input-id which meant that the cell was broadcasting wrong values
and could easily end up with an infinite loop. No good.

Now, the cell uses its own ID when doing a put after receiving inputs.

As well, I thought there might be an issue with doubling up the
watcher loops, so I put a guard in to prevent the user from
accidentally starting many watcher loops without realizing it.